+++ src/v2.x/packages/shared/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @copilotkitnext/shared
 
+## 1.51.4-next.1
+
+## 1.51.4-next.0
+
+## 1.51.3
+
+### Patch Changes
+
+- d36fc1e: Add UMD export
+
+## 1.51.3-next.8
+
+### Patch Changes
+
+- d36fc1e: Add UMD export
+
+## 1.51.3-next.7
+
+## 1.51.3-next.6
+
+## 1.51.3-next.5
+
+## 1.51.3-next.4
+
 ## 1.51.3-next.3
 
 ## 1.51.3-next.2